Update T265 SLAM configuration and logging paths
- Changed base path for logging in T265Config from "t265" to "/world/ego".
- Updated paths for left and right pinhole video streams and fisheye images to reflect the new base path.
- Removed unused pose path variable.
- Adjusted logging calls to use the new paths for fisheye images.
- Modified StereoEgoSequence to accommodate new entity paths for video streams and camera entities.
- Removed redundant confidence-related functions and classes from rerun_log_utils.py, moving them to a new file rerun_custom_types.py for better organization.
- Updated read_h264_samples_from_rrd function to normalize video entity paths.
